fix: 소유자 검증 활성화

+++ src/main/java/com/team6/moduply/playlist/service/PlaylistService.java
@@ -1,7 +1,6 @@
 package com.team6.moduply.playlist.service;
 
 import com.team6.moduply.common.pagination.CursorResponse;
-import com.team6.moduply.content.entity.Content;
 import com.team6.moduply.content.repository.ContentRepository;
 import com.team6.moduply.playlist.dto.PlaylistCreateRequest;
 import com.team6.moduply.playlist.dto.PlaylistDto;
@@ -34,8 +33,7 @@ public class PlaylistService {
 
   @Transactional
   public PlaylistDto create(PlaylistCreateRequest request, UUID ownerId) {
-    // TODO: ownerId는 현재 임시로 파라미터로 받음
-    // TODO: 인증 담당자 작업 완료 후 @AuthenticationPrincipal 또는 SecurityContextHolder로 교체 필요
+
     Playlist playlist = Playlist.builder()
         .ownerId(ownerId)
         .title(request.title())
@@ -49,16 +47,17 @@ public class PlaylistService {
 
   @Transactional
   public PlaylistDto update(UUID playlistId, PlaylistUpdateRequest request, UUID ownerId) {
-    // TODO: 인증 담당자 작업 완료 후 ownerId 교체 필요
+
     Playlist playlist = playlistRepository.findById(playlistId)
         .orElseThrow(() -> new PlaylistException(
             PlaylistErrorCode.PLAYLIST_NOT_FOUND,
             Map.of("playlistId", playlistId)
         ));
-// TODO: 인증 연동 후 소유자 검증 활성화 필요
-//    if (!playlist.getOwnerId().equals(ownerId)) {
-//      throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, playlistId);
-//    }
+
+    if (!playlist.getOwnerId().equals(ownerId)) {
+      throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN,
+          Map.of("playlistId", playlistId));
+    }
 
     playlist.update(request.title(), request.description());
 
@@ -67,17 +66,16 @@ public class PlaylistService {
 
   @Transactional
   public void delete(UUID playlistId, UUID ownerId) {
-    // TODO: 인증 담당자 작업 완료 후 ownerId 교체 필요
     Playlist playlist = playlistRepository.findById(playlistId)
         .orElseThrow(() -> new PlaylistException(
             PlaylistErrorCode.PLAYLIST_NOT_FOUND,
             Map.of("playlistId", playlistId)
         ));
 
-// TODO: 인증 연동 후 소유자 검증 활성화 필요
-//    if (!playlist.getOwnerId().equals(ownerId)) {
-//      throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, playlistId);
-//    }
+    if (!playlist.getOwnerId().equals(ownerId)) {
+      throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN,
+          Map.of("playlistId", playlistId));
+    }
 
     playlistRepository.delete(playlist);
   }
@@ -132,17 +130,15 @@ public class PlaylistService {
 
   @Transactional
   public void addContent(UUID playlistId, UUID contentId, UUID ownerId) {
-    // TODO: 인증 담당자 작업 완료 후 ownerId 교체 필요
     Playlist playlist = playlistRepository.findById(playlistId)
         .orElseThrow(() -> new PlaylistException(
             PlaylistErrorCode.PLAYLIST_NOT_FOUND,
             Map.of("playlistId", playlistId)
         ));
 
-    // TODO: 인증 연동 후 소유자 검증 활성화 필요
-    // if (!playlist.getOwnerId().equals(ownerId)) {
-    //     throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, Map.of("playlistId", playlistId));
-    // }
+     if (!playlist.getOwnerId().equals(ownerId)) {
+         throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, Map.of("playlistId", playlistId));
+     }
 
     if (!contentRepository.existsById(contentId)) {
       throw new PlaylistException(
@@ -170,17 +166,15 @@ public class PlaylistService {
 
   @Transactional
   public void removeContent(UUID playlistId, UUID contentId, UUID ownerId) {
-    // TODO: 인증 담당자 작업 완료 후 ownerId 교체 필요
     Playlist playlist = playlistRepository.findById(playlistId)
         .orElseThrow(() -> new PlaylistException(
             PlaylistErrorCode.PLAYLIST_NOT_FOUND,
             Map.of("playlistId", playlistId)
         ));
 
-    // TODO: 인증 연동 후 소유자 검증 활성화 필요
-    // if (!playlist.getOwnerId().equals(ownerId)) {
-    //     throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, Map.of("playlistId", playlistId));
-    // }
+     if (!playlist.getOwnerId().equals(ownerId)) {
+         throw new PlaylistException(PlaylistErrorCode.PLAYLIST_FORBIDDEN, Map.of("playlistId", playlistId));
+     }
 
     PlaylistContent playlistContent = playlistContentRepository
         .findByPlaylistAndContentId(playlist, contentId)
